feat: Work Requirement Service v2 delta — RowVersion concurrency, per-line visibility/audience filtering, search + audit
Implements docs/API/specs/update_requirement-schema.sql (v2-delta-summary.docx
§3): RequirementTemplateVersion.RowVersion concurrency guard with required
If-Match on PUT .../requirements; new requirement.RequirementLineVisibility
table + VisibilityLevel enum wired through resolve/revise so every requirement
line can be marked Customer/Provider/Internal; ?audience= filtering on
GetResolved/Explain and ?customerVisibleOnly= on Compare; new GET
.../search and GET .../{id}/audit endpoints. Adds migration
AddWorkRequirementVisibilityAndConcurrency and the first OpenAPI coverage
this reconciled surface has had (6 touched endpoints).

Live Testcontainers testing found and fixed a real interop bug: the server
parsed the raw If-Match header with long.TryParse, which rejects a correctly
RFC-7232-quoted ETag ("1") that real HTTP clients send.

74/74 unit tests passing; WorkRequirementsApiTests 7/7 passing individually
against real Postgres.

+++ src/Eswmp.Work/Controllers/WorkRequirementErrors.cs
@@ -34,7 +34,7 @@ internal static class WorkRequirementControllerExtensions
         controller.ErrorResult(StatusCodes.Status409Conflict, "IDEMPOTENCY_CONFLICT",
             $"Idempotency-Key '{idempotencyKey}' was already used with a different request body.");
 
-    public static ObjectResult VersionConflict(this ControllerBase controller, int expectedVersion, int? currentVersion) =>
+    public static ObjectResult VersionConflict(this ControllerBase controller, long expectedVersion, long? currentVersion) =>
         controller.ErrorResult(StatusCodes.Status412PreconditionFailed, "VERSION_CONFLICT",
             currentVersion is null
                 ? $"Expected version {expectedVersion} no longer matches the current version."
